Move StudentEnrollRequest to its own file
